reorganize the filter package; moved Clause -> CustomFilter (left aliases to Clause) and also moved functionality to separate files and a builder package for the Filter functionality
